arm_tests_fix

+++ crypto/src/dap_enc_sphincsplus.c
@@ -36,7 +36,7 @@ void dap_enc_sig_sphincsplus_key_new_generate(dap_enc_key_t *a_key, const void *
     }
     
     // seed norming
-    uint64_t l_key_size = dap_enc_sphincsplus_crypto_sign_seedbytes(s_default_config);
+    uint32_t l_key_size = dap_enc_sphincsplus_crypto_sign_seedbytes(s_default_config);
     unsigned char l_seedbuf[l_key_size];
 
     if(a_seed && a_seed_size > 0) {
@@ -99,7 +99,7 @@ size_t dap_enc_sig_sphincsplus_get_sign_msg(dap_enc_key_t *a_key, const void *a_
 {
 
     sphincsplus_private_key_t *l_skey = a_key->priv_key_data;
-    uint64_t l_sign_bytes = dap_enc_sphincsplus_crypto_sign_bytes(&l_skey->params);
+    uint32_t l_sign_bytes = dap_enc_sphincsplus_crypto_sign_bytes(&l_skey->params);
     
     if(a_out_size_max < l_sign_bytes) {
         log_it(L_ERROR, "Bad signature size");
@@ -151,7 +151,7 @@ size_t dap_enc_sig_sphincsplus_open_sign_msg(dap_enc_key_t *a_key, const void *a
         const size_t a_out_size_max)
 {
     sphincsplus_public_key_t *l_pkey = a_key->pub_key_data;
-    uint64_t l_sign_bytes = dap_enc_sphincsplus_crypto_sign_bytes(&l_pkey->params);
+    uint32_t l_sign_bytes = dap_enc_sphincsplus_crypto_sign_bytes(&l_pkey->params);
 
     if(a_out_size_max < l_sign_bytes) {
         log_it(L_ERROR, "Bad signature size");
@@ -169,7 +169,7 @@ size_t dap_enc_sig_sphincsplus_open_sign_msg(dap_enc_key_t *a_key, const void *a
             pthread_mutex_unlock(&s_sign_mtx);
             return 0;
         }
-        uint64_t l_res_size = 0;
+        uint32_t l_res_size = 0;
         if (sphincsplus_crypto_sign_open(a_msg_out, &l_res_size, l_sign->sig_data, l_sign->sig_len, l_pkey->data))
             log_it(L_ERROR, "Failed to verify signature");
     pthread_mutex_unlock(&s_sign_mtx);
@@ -194,10 +194,10 @@ uint8_t *dap_enc_sphincsplus_write_private_key(const void *a_private_key, size_t
     dap_return_val_if_pass(!a_private_key, NULL);
     sphincsplus_private_key_t *l_private_key = (sphincsplus_private_key_t *)a_private_key;
 // func work
-    uint64_t l_secret_length = dap_enc_sphincsplus_crypto_sign_secretkeybytes(&l_private_key->params);
-    uint64_t l_buflen = sizeof(uint64_t) + sizeof(sphincsplus_base_params_t) + l_secret_length;
+    uint32_t l_secret_length = dap_enc_sphincsplus_crypto_sign_secretkeybytes(&l_private_key->params);
+    uint32_t l_buflen = sizeof(uint32_t) + sizeof(sphincsplus_base_params_t) + l_secret_length;
     uint8_t *l_buf = dap_serialize_multy(NULL, l_buflen, 6,
-        &l_buflen, sizeof(uint64_t),
+        &l_buflen, sizeof(uint32_t),
         &l_private_key->params, sizeof(sphincsplus_base_params_t),
         l_private_key->data, l_secret_length
     );
@@ -210,22 +210,22 @@ uint8_t *dap_enc_sphincsplus_write_private_key(const void *a_private_key, size_t
 sphincsplus_private_key_t *dap_enc_sphincsplus_read_private_key(const uint8_t *a_buf, size_t a_buflen)
 {
 // in work
-    dap_return_val_if_pass(!a_buf || a_buflen < sizeof(uint64_t) + sizeof(sphincsplus_base_params_t), NULL);
+    dap_return_val_if_pass(!a_buf || a_buflen < sizeof(uint32_t) + sizeof(sphincsplus_base_params_t), NULL);
 // func work
-    uint64_t l_buflen = 0;
-    uint64_t l_skey_len = a_buflen -  sizeof(uint64_t) - sizeof(sphincsplus_base_params_t);
+    uint32_t l_buflen = 0;
+    uint32_t l_skey_len = a_buflen -  sizeof(uint32_t) - sizeof(sphincsplus_base_params_t);
 
     sphincsplus_private_key_t *l_skey = NULL;
     DAP_NEW_Z_RET_VAL(l_skey, sphincsplus_private_key_t, NULL, NULL);
     DAP_NEW_Z_SIZE_RET_VAL(l_skey->data, uint8_t, l_skey_len, NULL, l_skey);
 
     int l_res_des = dap_deserialize_multy(a_buf, a_buflen, 6, 
-        &l_buflen, sizeof(uint64_t),
+        &l_buflen, sizeof(uint32_t),
         &l_skey->params, sizeof(sphincsplus_base_params_t),
         l_skey->data, l_skey_len
     );
 // out work
-    uint64_t l_skey_len_exp = dap_enc_sphincsplus_crypto_sign_secretkeybytes(&l_skey->params);
+    uint32_t l_skey_len_exp = dap_enc_sphincsplus_crypto_sign_secretkeybytes(&l_skey->params);
     if (l_res_des) {
         log_it(L_ERROR,"::read_private_key() deserialise public key, err code %d", l_res_des);
         DAP_DEL_MULTY(l_skey->data, l_skey);
@@ -247,10 +247,10 @@ uint8_t *dap_enc_sphincsplus_write_public_key(const void* a_public_key, size_t *
     dap_return_val_if_pass(!a_public_key, NULL);
     sphincsplus_public_key_t *l_public_key = (sphincsplus_public_key_t *)a_public_key;
 // func work
-    uint64_t l_public_length = dap_enc_sphincsplus_crypto_sign_publickeybytes(&l_public_key->params);
-    uint64_t l_buflen = sizeof(uint64_t) + sizeof(sphincsplus_base_params_t) + l_public_length;
+    uint32_t l_public_length = dap_enc_sphincsplus_crypto_sign_publickeybytes(&l_public_key->params);
+    uint32_t l_buflen = sizeof(uint32_t) + sizeof(sphincsplus_base_params_t) + l_public_length;
     uint8_t *l_buf = dap_serialize_multy(NULL, l_buflen, 6, 
-        &l_buflen, sizeof(uint64_t),
+        &l_buflen, sizeof(uint32_t),
         &l_public_key->params, sizeof(sphincsplus_base_params_t),
         l_public_key->data, l_public_length
     );
@@ -263,22 +263,22 @@ uint8_t *dap_enc_sphincsplus_write_public_key(const void* a_public_key, size_t *
 sphincsplus_public_key_t *dap_enc_sphincsplus_read_public_key(const uint8_t *a_buf, size_t a_buflen)
 {
 // in work
-    dap_return_val_if_pass(!a_buf || a_buflen < sizeof(uint64_t) + sizeof(sphincsplus_base_params_t), NULL);
+    dap_return_val_if_pass(!a_buf || a_buflen < sizeof(uint32_t) + sizeof(sphincsplus_base_params_t), NULL);
 // func work
-    uint64_t l_buflen = 0;
-    uint64_t l_pkey_len = a_buflen -  sizeof(uint64_t) - sizeof(sphincsplus_base_params_t);
+    uint32_t l_buflen = 0;
+    uint32_t l_pkey_len = a_buflen -  sizeof(uint32_t) - sizeof(sphincsplus_base_params_t);
 
     sphincsplus_public_key_t *l_pkey = NULL;
     DAP_NEW_Z_RET_VAL(l_pkey, sphincsplus_public_key_t, NULL, NULL);
     DAP_NEW_Z_SIZE_RET_VAL(l_pkey->data, uint8_t, l_pkey_len, NULL, l_pkey);
 
     int l_res_des = dap_deserialize_multy(a_buf, a_buflen, 6, 
-        &l_buflen, sizeof(uint64_t),
+        &l_buflen, sizeof(uint32_t),
         &l_pkey->params, sizeof(sphincsplus_base_params_t),
         l_pkey->data, l_pkey_len
     );
 // out work
-    uint64_t l_pkey_len_exp = dap_enc_sphincsplus_crypto_sign_publickeybytes(&l_pkey->params);
+    uint32_t l_pkey_len_exp = dap_enc_sphincsplus_crypto_sign_publickeybytes(&l_pkey->params);
     if (l_res_des) {
         log_it(L_ERROR,"::read_public_key() deserialise public key, err code %d", l_res_des);
         DAP_DEL_MULTY(l_pkey->data, l_pkey);
@@ -300,11 +300,11 @@ uint8_t *dap_enc_sphincsplus_write_signature(const void *a_sign, size_t *a_bufle
     dap_return_val_if_pass(!a_sign, NULL);
     sphincsplus_signature_t *l_sign = (sphincsplus_signature_t *)a_sign;
 // func work
-    uint64_t l_buflen = l_sign->sig_len + sizeof(uint64_t) * 2 + sizeof(sphincsplus_base_params_t);
+    uint32_t l_buflen = l_sign->sig_len + sizeof(uint32_t) * 2 + sizeof(sphincsplus_base_params_t);
     uint8_t *l_buf = dap_serialize_multy(NULL, l_buflen, 8, 
-        &l_buflen, sizeof(uint64_t),
+        &l_buflen, sizeof(uint32_t),
         &l_sign->sig_params, sizeof(sphincsplus_base_params_t),
-        &l_sign->sig_len, sizeof(uint64_t),
+        &l_sign->sig_len, sizeof(uint32_t),
         l_sign->sig_data, l_sign->sig_len
     );
 // out work
@@ -316,18 +316,18 @@ uint8_t *dap_enc_sphincsplus_write_signature(const void *a_sign, size_t *a_bufle
 sphincsplus_signature_t *dap_enc_sphincsplus_read_signature(const uint8_t *a_buf, size_t a_buflen)
 {
 // in work
-    dap_return_val_if_pass(!a_buf || a_buflen < sizeof(uint64_t) * 2 + sizeof(sphincsplus_base_params_t), NULL);
+    dap_return_val_if_pass(!a_buf || a_buflen < sizeof(uint32_t) * 2 + sizeof(sphincsplus_base_params_t), NULL);
 // func work
-    uint64_t l_buflen;
-    uint64_t l_sig_len = a_buflen - sizeof(uint64_t) * 2 - sizeof(sphincsplus_base_params_t);
+    uint32_t l_buflen;
+    uint32_t l_sig_len = a_buflen - sizeof(uint32_t) * 2 - sizeof(sphincsplus_base_params_t);
     sphincsplus_signature_t* l_sign = NULL;
     DAP_NEW_Z_RET_VAL(l_sign, sphincsplus_signature_t, NULL, NULL);
     DAP_NEW_Z_SIZE_RET_VAL(l_sign->sig_data, uint8_t, l_sig_len, NULL, l_sign);
 
     int l_res_des = dap_deserialize_multy(a_buf, a_buflen, 8, 
-        &l_buflen, sizeof(uint64_t),
+        &l_buflen, sizeof(uint32_t),
         &l_sign->sig_params, sizeof(sphincsplus_base_params_t),
-        &l_sign->sig_len, sizeof(uint64_t),
+        &l_sign->sig_len, sizeof(uint32_t),
         l_sign->sig_data, l_sig_len
     );
 // out work
@@ -369,14 +369,14 @@ void sphincsplus_signature_delete(sphincsplus_signature_t *a_sig){
 /*
  * Returns the length of a secret key, in bytes
  */
-uint64_t dap_enc_sphincsplus_crypto_sign_secretkeybytes(const sphincsplus_base_params_t *a_params)
+uint32_t dap_enc_sphincsplus_crypto_sign_secretkeybytes(const sphincsplus_base_params_t *a_params)
 {
     pthread_mutex_lock(&s_sign_mtx);
         if (sphincsplus_set_params(a_params)) {
             pthread_mutex_unlock(&s_sign_mtx);
             return 0;
         }
-        uint64_t l_ret = sphincsplus_crypto_sign_secretkeybytes();
+        uint32_t l_ret = sphincsplus_crypto_sign_secretkeybytes();
     pthread_mutex_unlock(&s_sign_mtx);
     return l_ret;
 }
@@ -384,14 +384,14 @@ uint64_t dap_enc_sphincsplus_crypto_sign_secretkeybytes(const sphincsplus_base_p
 /*
  * Returns the length of a public key, in bytes
  */
-uint64_t dap_enc_sphincsplus_crypto_sign_publickeybytes(const sphincsplus_base_params_t *a_params)
+uint32_t dap_enc_sphincsplus_crypto_sign_publickeybytes(const sphincsplus_base_params_t *a_params)
 {
     pthread_mutex_lock(&s_sign_mtx);
         if (sphincsplus_set_params(a_params)) {
             pthread_mutex_unlock(&s_sign_mtx);
             return 0;
         }
-        uint64_t l_ret = sphincsplus_crypto_sign_publickeybytes();
+        uint32_t l_ret = sphincsplus_crypto_sign_publickeybytes();
     pthread_mutex_unlock(&s_sign_mtx);
     return l_ret;
 }
@@ -399,14 +399,14 @@ uint64_t dap_enc_sphincsplus_crypto_sign_publickeybytes(const sphincsplus_base_p
 /*
  * Returns the length of the seed required to generate a key pair, in bytes
  */
-uint64_t dap_enc_sphincsplus_crypto_sign_seedbytes(sphincsplus_config_t a_config)
+uint32_t dap_enc_sphincsplus_crypto_sign_seedbytes(sphincsplus_config_t a_config)
 {
     pthread_mutex_lock(&s_sign_mtx);
         if (sphincsplus_set_config(a_config)) {
             pthread_mutex_unlock(&s_sign_mtx);
             return 0;
         }
-        uint64_t l_ret = sphincsplus_crypto_sign_seedbytes();
+        uint32_t l_ret = sphincsplus_crypto_sign_seedbytes();
     pthread_mutex_unlock(&s_sign_mtx);
     return l_ret;
 }
@@ -414,19 +414,19 @@ uint64_t dap_enc_sphincsplus_crypto_sign_seedbytes(sphincsplus_config_t a_config
 /*
  * Returns the length of a signature, in bytes
  */
-uint64_t dap_enc_sphincsplus_crypto_sign_bytes(const sphincsplus_base_params_t *a_params)
+uint32_t dap_enc_sphincsplus_crypto_sign_bytes(const sphincsplus_base_params_t *a_params)
 {
     pthread_mutex_lock(&s_sign_mtx);
         if (sphincsplus_set_params(a_params)) {
             pthread_mutex_unlock(&s_sign_mtx);
             return 0;
         }
-        uint64_t l_ret = sphincsplus_crypto_sign_bytes();
+        uint32_t l_ret = sphincsplus_crypto_sign_bytes();
     pthread_mutex_unlock(&s_sign_mtx);
     return l_ret;
 }
 
-uint64_t dap_enc_sphincsplus_calc_signature_unserialized_size()
+uint32_t dap_enc_sphincsplus_calc_signature_unserialized_size()
 {
     return sizeof(sphincsplus_signature_t); 
 }
